+++ test.js
@@ -193,54 +193,3 @@ for (const [name, linkify] of Object.entries({
 	});
 }
 
-test('LinkifyUrls component: basic functionality', async t => {
-	// Test that the component function exists
-	const {LinkifyUrls} = await import('./index.js');
-	t.is(typeof LinkifyUrls, 'function');
-
-	// Mock React globally
-	const originalReact = globalThis.React;
-	globalThis.React = {
-		useRef() {
-			return {current: undefined};
-		},
-		useEffect() {},
-		useMemo(factory) {
-			return factory();
-		},
-		createElement: (type, properties, ...children) => ({type, properties, children}),
-		Fragment: Symbol('React.Fragment'),
-	};
-
-	try {
-		// Test with non-string children (should return as-is)
-		// eslint-disable-next-line new-cap
-		const result = LinkifyUrls({children: null});
-		t.is(result, null);
-
-		// Test with string children (should return React element)
-		// eslint-disable-next-line new-cap
-		const stringResult = LinkifyUrls({children: 'test'});
-		t.is(stringResult.type, globalThis.React.Fragment);
-	} finally {
-		// Restore original React
-		globalThis.React = originalReact;
-	}
-});
-
-test('LinkifyUrls component: throws when React not available', async t => {
-	const {LinkifyUrls} = await import('./index.js');
-
-	// Ensure React is not available
-	const originalReact = globalThis.React;
-	delete globalThis.React;
-
-	try {
-		// eslint-disable-next-line new-cap
-		const error = t.throws(() => LinkifyUrls({children: 'test'}));
-		t.is(error.message, 'LinkifyUrls requires React to be available globally. Make sure React is imported in your app.');
-	} finally {
-		// Restore original React
-		globalThis.React = originalReact;
-	}
-});
